requirements.txt
absl-py==1.4.0
adapter-transformers==3.2.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.8
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
bitarray==2.7.3
bs4==0.0.1
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.0
colorama==0.4.6
Cython==0.29.34
datasets==2.10.1
dill==0.3.6
evaluate==0.4.0
faiss-gpu==1.7.2
filelock==3.10.0
frozenlist==1.3.3
fsspec==2023.3.0
huggingface-hub==0.13.2
hydra-core==1.0.7
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonlines==3.1.0
lit==15.0.7
lxml==4.9.2
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.0
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.0.6
packaging==23.0
pandas==1.5.3
portalocker==2.7.0
prettytable==3.7.0
prettyTables==1.1.5
protobuf==3.20.3
pyarrow==11.0.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
responses==0.18.0
rouge==1.0.1
rouge-score==0.1.2
sacrebleu==2.3.1
sentencepiece==0.1.99
six==1.16.0
soupsieve==2.4
sympy==1.11.1
tabulate==0.9.0
tensorboardX==2.6
tokenizers==0.13.2
torch==1.9.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
